modify generics

+++ theories/generics.go
@@ -1,6 +1,6 @@
 package theories
 
-func Sum[K comparable, V int | float32](m map[K]V) V {
+func Sum[K comparable, V int | float64](m map[K]V) V {
 	var s V
 	for _, v := range m {
 		s += v
